Factor out common Question Generator component
